psu-ng: Change pgoodFault type to size_t
The Misc Guidelines section of the coding conventions indicate that we
should "Always use size_t or ssize_t for things that are sizes, counts,
etc. ...."

https: //github.com/openbmc/docs/blob/master/cpp-style-and-conventions.md

+++ phosphor-power-supply/power_supply.hpp
@@ -448,7 +448,7 @@ class PowerSupply
      *
      * Considered faulted if reaches DEGLITCH_LIMIT.
      */
-    int pgoodFault = 0;
+    size_t pgoodFault = 0;
 
     /**
      * @brief Power Supply Kill fault.
